Add support to logo images found in new mediatek devices using bgra8888 instead of rbg565.
To distinguish between logo versions there is a little difference in the header. I have found that versions using rgb565 contains the word "LOGO" in capital letters, while ones using bgra8888 use lowercase. I have tested it only in a few devices, so if there are exceptions to that rule the scripts will need to be corrected.